Ajustement droit modification ajout de post/catégories

+++ src/Controller/User/PostController.php
@@ -4,6 +4,7 @@ namespace App\Controller\User;
 
 use App\Entity\Post;
 use App\Form\PostType;
+use DateTimeImmutable;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -22,6 +23,8 @@ class PostController extends AbstractController
         $form = $this->createForm(PostType::class, $post);
         $form->handleRequest($request);
         if ($form->isSubmitted() && $form->isValid()) {
+            $post->setUser($this->getUser());
+            $post->setCreatedAt(new DateTimeImmutable());
             $entityManager->persist($post);
             $entityManager->flush();
             return $this->redirectToRoute('app_post_index', [], Response::HTTP_SEE_OTHER);
@@ -35,6 +38,9 @@ class PostController extends AbstractController
     #[Route('/{id}/edit', name: 'app_post_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Post $post, EntityManagerInterface $entityManager): Response
     {
+        if ($post->getUser() != $this->getUser() && !$this->isGranted("ROLE_ADMIN")) {
+            return $this->redirectToRoute('app_post_index', [], Response::HTTP_SEE_OTHER);
+        }
         $form = $this->createForm(PostType::class, $post);
         $form->handleRequest($request);
         if ($form->isSubmitted() && $form->isValid()) {
@@ -50,6 +56,9 @@ class PostController extends AbstractController
     #[Route('/{id}', name: 'app_post_delete', methods: ['POST'])]
     public function delete(Request $request, Post $post, EntityManagerInterface $entityManager): Response
     {
+        if ($post->getUser() != $this->getUser() && !$this->isGranted("ROLE_ADMIN")) {
+            return $this->redirectToRoute('app_post_index', [], Response::HTTP_SEE_OTHER);
+        }
         if ($this->isCsrfTokenValid('delete'.$post->getId(), $request->request->get('_token'))) {
             $entityManager->remove($post);
             $entityManager->flush();
